+++ appservice/src/tree/AppSettingsTreeItem.ts
@@ -5,10 +5,8 @@
 
 import { StringDictionary } from 'azure-arm-website/lib/models';
 import * as path from 'path';
-import { MessageItem } from 'vscode';
-import { AzureParentTreeItem, AzureTreeItem, DialogResponses } from 'vscode-azureextensionui';
+import { AzureParentTreeItem, AzureTreeItem } from 'vscode-azureextensionui';
 import { ext } from '../extensionVariables';
-import { localize } from '../localize';
 import { AppSettingTreeItem } from './AppSettingTreeItem';
 import { ISiteTreeRoot } from './ISiteTreeRoot';
 
@@ -120,77 +118,5 @@ export class AppSettingsTreeItem extends AzureParentTreeItem<ISiteTreeRoot> {
         }
 
         return <StringDictionary>this._settings;
-    }
-
-    public async confirmOverwriteSettings(sourceSettings: { [key: string]: string }, destinationSettings: { [key: string]: string }, destinationName: string): Promise<void> {
-        function logKey(key: string): void {
-            ext.outputChannel.appendLine(`- ${key}`);
-        }
-
-        let suppressPrompt: boolean = false;
-        let overwriteSetting: boolean = false;
-
-        const addedKeys: string[] = [];
-        const updatedKeys: string[] = [];
-        const userIgnoredKeys: string[] = [];
-        const matchingKeys: string[] = [];
-
-        for (const key of Object.keys(sourceSettings)) {
-            if (destinationSettings[key] === undefined) {
-                addedKeys.push(key);
-                destinationSettings[key] = sourceSettings[key];
-            } else if (destinationSettings[key] !== sourceSettings[key]) {
-                if (!suppressPrompt) {
-                    const yesToAll: MessageItem = { title: localize('yesToAll', 'Yes to all') };
-                    const noToAll: MessageItem = { title: localize('noToAll', 'No to all') };
-                    const message: string = localize('overwriteSetting', 'Setting "{0}" already exists in "{1}". Overwrite?', key, destinationName);
-                    const result: MessageItem = await ext.ui.showWarningMessage(message, { modal: true }, DialogResponses.yes, yesToAll, DialogResponses.no, noToAll);
-                    if (result === DialogResponses.yes) {
-                        overwriteSetting = true;
-                    } else if (result === yesToAll) {
-                        overwriteSetting = true;
-                        suppressPrompt = true;
-                    } else if (result === DialogResponses.no) {
-                        overwriteSetting = false;
-                    } else if (result === noToAll) {
-                        overwriteSetting = false;
-                        suppressPrompt = true;
-                    }
-                }
-
-                if (overwriteSetting) {
-                    updatedKeys.push(key);
-                    destinationSettings[key] = sourceSettings[key];
-                } else {
-                    userIgnoredKeys.push(key);
-                }
-            } else {
-                matchingKeys.push(key);
-            }
-        }
-
-        if (addedKeys.length > 0) {
-            ext.outputChannel.appendLine(localize('addedKeys', 'Added the following settings:'));
-            addedKeys.forEach(logKey);
-        }
-
-        if (updatedKeys.length > 0) {
-            ext.outputChannel.appendLine(localize('updatedKeys', 'Updated the following settings:'));
-            updatedKeys.forEach(logKey);
-        }
-
-        if (matchingKeys.length > 0) {
-            ext.outputChannel.appendLine(localize('matchingKeys', 'Ignored the following settings that were already the same:'));
-            matchingKeys.forEach(logKey);
-        }
-
-        if (userIgnoredKeys.length > 0) {
-            ext.outputChannel.appendLine(localize('userIgnoredKeys', 'Ignored the following settings based on user input:'));
-            userIgnoredKeys.forEach(logKey);
-        }
-
-        if (Object.keys(destinationSettings).length > Object.keys(sourceSettings).length) {
-            ext.outputChannel.appendLine(localize('noDeleteKey', 'WARNING: This operation will not delete any settings in "{0}". You must manually delete settings if desired.', destinationName));
-        }
     }
 }
